Authorization and centralized error handling

+++ src/services/post.service.js
@@ -73,24 +73,39 @@ export const createPost = async (postData, authorId) => {
   }
 };
 
-export const updatePost = async (id, postData) => {
+export const updatePost = async (id, postData, userId) => {
   const { title, content } = postData;
-  const [result] = await pool.query(
-    "UPDATE posts SET title = ?, content = ? WHERE id = ?",
-    [title, content, id]
-  );
-  if (result.affectedRows === 0) {
-    throw new ApiError(404, "Post not found.");
+
+  const post = await getPostById(id);
+
+  if (post.authorId !== userId) {
+    throw new ApiError(
+      403,
+      "Forbidden: You do not have permission to edit this post."
+    );
   }
-  return getPostById(id);
+
+  await pool.query("UPDATE posts SET title = ?, content = ? WHERE id = ?", [
+    title,
+    content,
+    id,
+  ]);
+  const updatePost = await getPostById(id);
+  return updatePost;
 };
 
-export const deletePost = async (id) => {
-  const [result] = await pool.query("DELETE FROM posts WHERE id = ?", [id]);
-  if (result.affectedRows === 0) {
-    throw new ApiError(404, "Post not found.");
+export const deletePost = async (id, userId) => {
+  const post = await getPostById(id);
+
+  if (post.authorId !== userId) {
+    throw new ApiError(
+      403,
+      "Forbidden: You do not have permission to delete this post."
+    );
   }
-  return true;
+
+  const [result] = await pool.query("DELETE FROM posts WHERE id = ?", [id]);
+  return result.affectedRows;
 };
 
 export const partiallyUpdatePost = async (id, updates) => {
